separated new test case to its own file, added neg. case
